# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Computer API",
        "description": "Patrick Simple Computer API",
        "contact": {
            "email": "n00201161@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/computers/{id}": {
            "get": {
                "tags": [
                    "Computers"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets a computer by ID",
                "operationId": "1244ba519c2f0a0843f22a23dd4f6134",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Computer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Computers"
                ],
                "summary": "Update an existing Computer",
                "description": "Updates the computer in the DB",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Computer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "brand",
                                    "graphics_card",
                                    "processor",
                                    "storage",
                                    "ram",
                                    "price",
                                    "brand_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this computer"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "graphics_card": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "processor": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "storage": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "ram": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Me"
                                    },
                                    "developers": {
                                        "type": "json",
                                        "format": "json",
                                        "example": [
                                            1,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Computers"
                ],
                "summary": "Delete a Computer",
                "description": "Delete Computer",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Computer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/developers": {
            "get": {
                "tags": [
                    "Developers"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the developers",
                "operationId": "5d4d2fcb34324751c78c26412bbfc10c",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Developers in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/developers/{id}": {
            "get": {
                "tags": [
                    "Developers"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets a developer by ID",
                "operationId": "e1d46842f7196b812b00df3e26c99fac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Developer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}